tests/manual/requirements.txt
json5
python-dotenv>=1.0.0
